refactor virtual farmbot
