+++ webpack/farm_designer/map/virtual_farmbot/__tests__/index_test.tsx
@@ -1,0 +1,55 @@
+const mockStorj: Dictionary<boolean> = {};
+
+jest.mock("../../../../session", () => {
+  return {
+    Session: {
+      getBool: (k: string) => {
+        mockStorj[k] = !!mockStorj[k];
+        return mockStorj[k];
+      },
+    }
+  };
+});
+
+import * as React from "react";
+import { VirtualFarmBot } from "../index";
+import { shallow } from "enzyme";
+import { VirtualFarmBotProps } from "../../interfaces";
+import { Dictionary } from "farmbot";
+
+describe("<VirtualFarmBot/>", () => {
+  function fakeProps(): VirtualFarmBotProps {
+    return {
+      botLocationData: {
+        position: { x: 0, y: 0, z: 0 },
+        scaled_encoders: { x: undefined, y: undefined, z: undefined },
+        raw_encoders: { x: undefined, y: undefined, z: undefined },
+      },
+      mapTransformProps: {
+        quadrant: 1, gridSize: { x: 3000, y: 1500 }
+      },
+      plantAreaOffset: { x: 100, y: 100 }
+    };
+  }
+
+  it("shows bot position", () => {
+    const wrapper = shallow(<VirtualFarmBot {...fakeProps() } />);
+    const figures = wrapper.find("BotFigure");
+    expect(figures.length).toEqual(1);
+    expect(figures.last().props().name).toEqual("motor-position");
+  });
+
+  it("shows trail", () => {
+    mockStorj["displayTrail"] = true;
+    const wrapper = shallow(<VirtualFarmBot {...fakeProps() } />);
+    expect(wrapper.find("BotTrail").length).toEqual(1);
+  });
+
+  it("shows encoder position", () => {
+    mockStorj["encoderFigure"] = true;
+    const wrapper = shallow(<VirtualFarmBot {...fakeProps() } />);
+    const figures = wrapper.find("BotFigure");
+    expect(figures.length).toEqual(2);
+    expect(figures.last().props().name).toEqual("encoder-position");
+  });
+});
